Refactor util functions and add in a generic function to find file types in a dir
